feat: implement sign transaction APDU command handler and UI flow
This commit introduces the sign transaction functionality with several key improvements:
- Implement handleSignTransaction function with buffer-based input and chunked data support
- Update dispatcher to support INS_SIGN_TRANSACTION command with extended buffer handling
- Add transaction validation and UI confirmation flow
- Modify send_response helper to return int for sign transaction response
- Add crypto_sign_transaction function for transaction signing
- Update UI display logic to support transaction signing
- Refactor sign transaction handler to use new buffer and error handling approach

+++ src/helper/send_response.h
@@ -5,5 +5,5 @@
 
 int helper_send_response_public_key();
 int helper_send_response_address();
-uint8_t helper_send_response_sign_transaction();
+int helper_send_response_sign_transaction();
 int helper_send_response_sign();
